Do not let `Permalink` refer to its subclass `PeepholePermalink` in its static initializer
