+++ core/src/main/java/hudson/model/Job.java
@@ -92,6 +92,7 @@ import jenkins.model.DirectlyModifiableTopLevelItemGroup;
 import jenkins.model.Jenkins;
 import jenkins.model.JenkinsLocationConfiguration;
 import jenkins.model.ModelObjectWithChildren;
+import jenkins.model.PeepholePermalink;
 import jenkins.model.ProjectNamingStrategy;
 import jenkins.model.RunIdMigrator;
 import jenkins.model.lazy.LazyBuildMixIn;
@@ -941,7 +942,7 @@ public abstract class Job<JobT extends Job<JobT, RunT>, RunT extends Run<JobT, R
     @Exported
     @QuickSilver
     public RunT getLastSuccessfulBuild() {
-        return (RunT) Permalink.LAST_SUCCESSFUL_BUILD.resolve(this);
+        return (RunT) PeepholePermalink.LAST_SUCCESSFUL_BUILD.resolve(this);
     }
 
     /**
@@ -951,7 +952,7 @@ public abstract class Job<JobT extends Job<JobT, RunT>, RunT extends Run<JobT, R
     @Exported
     @QuickSilver
     public RunT getLastUnsuccessfulBuild() {
-        return (RunT) Permalink.LAST_UNSUCCESSFUL_BUILD.resolve(this);
+        return (RunT) PeepholePermalink.LAST_UNSUCCESSFUL_BUILD.resolve(this);
     }
 
     /**
@@ -961,7 +962,7 @@ public abstract class Job<JobT extends Job<JobT, RunT>, RunT extends Run<JobT, R
     @Exported
     @QuickSilver
     public RunT getLastUnstableBuild() {
-        return (RunT) Permalink.LAST_UNSTABLE_BUILD.resolve(this);
+        return (RunT) PeepholePermalink.LAST_UNSTABLE_BUILD.resolve(this);
     }
 
     /**
@@ -971,7 +972,7 @@ public abstract class Job<JobT extends Job<JobT, RunT>, RunT extends Run<JobT, R
     @Exported
     @QuickSilver
     public RunT getLastStableBuild() {
-        return (RunT) Permalink.LAST_STABLE_BUILD.resolve(this);
+        return (RunT) PeepholePermalink.LAST_STABLE_BUILD.resolve(this);
     }
 
     /**
@@ -980,7 +981,7 @@ public abstract class Job<JobT extends Job<JobT, RunT>, RunT extends Run<JobT, R
     @Exported
     @QuickSilver
     public RunT getLastFailedBuild() {
-        return (RunT) Permalink.LAST_FAILED_BUILD.resolve(this);
+        return (RunT) PeepholePermalink.LAST_FAILED_BUILD.resolve(this);
     }
 
     /**
@@ -989,7 +990,7 @@ public abstract class Job<JobT extends Job<JobT, RunT>, RunT extends Run<JobT, R
     @Exported
     @QuickSilver
     public RunT getLastCompletedBuild() {
-        return (RunT) Permalink.LAST_COMPLETED_BUILD.resolve(this);
+        return (RunT) PeepholePermalink.LAST_COMPLETED_BUILD.resolve(this);
     }
 
     /**
@@ -1059,6 +1060,10 @@ public abstract class Job<JobT extends Job<JobT, RunT>, RunT extends Run<JobT, R
         if (totalDuration == 0) return -1;
 
         return Math.round((double) totalDuration / builds.size());
+    }
+
+    static {
+        PeepholePermalink.initialized();
     }
 
     /**
