+++ core/src/main/java/hudson/model/PermalinkProjectAction.java
@@ -25,7 +25,6 @@
 package hudson.model;
 
 import edu.umd.cs.findbugs.annotations.CheckForNull;
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.util.List;
 import java.util.concurrent.CopyOnWriteArrayList;
 import jenkins.model.PeepholePermalink;
@@ -111,114 +110,33 @@ public interface PermalinkProjectAction extends Action {
                 return job.getLastBuild();
             }
         };
-        public static final Permalink LAST_STABLE_BUILD = new PeepholePermalink() {
-            @Override
-            public String getDisplayName() {
-                return Messages.Permalink_LastStableBuild();
-            }
 
-            @Override
-            public String getId() {
-                return "lastStableBuild";
-            }
+        /** @deprecated use {@link PeepholePermalink#LAST_STABLE_BUILD} */
+        @Deprecated
+        public static Permalink LAST_STABLE_BUILD;
 
-            @Override
-            public boolean apply(Run<?, ?> run) {
-                return !run.isBuilding() && run.getResult() == Result.SUCCESS;
-            }
-        };
-        public static final Permalink LAST_SUCCESSFUL_BUILD = new PeepholePermalink() {
-            @Override
-            public String getDisplayName() {
-                return Messages.Permalink_LastSuccessfulBuild();
-            }
+        /** @deprecated use {@link PeepholePermalink#LAST_SUCCESSFUL_BUILD} */
+        @Deprecated
+        public static Permalink LAST_SUCCESSFUL_BUILD;
 
-            @Override
-            public String getId() {
-                return "lastSuccessfulBuild";
-            }
+        /** @deprecated use {@link PeepholePermalink#LAST_FAILED_BUILD} */
+        @Deprecated
+        public static Permalink LAST_FAILED_BUILD;
 
-            @SuppressFBWarnings(value = "NP_NULL_ON_SOME_PATH_FROM_RETURN_VALUE", justification = "TODO needs triage")
-            @Override
-            public boolean apply(Run<?, ?> run) {
-                return !run.isBuilding() && run.getResult().isBetterOrEqualTo(Result.UNSTABLE);
-            }
-        };
-        public static final Permalink LAST_FAILED_BUILD = new PeepholePermalink() {
-            @Override
-            public String getDisplayName() {
-                return Messages.Permalink_LastFailedBuild();
-            }
+        /** @deprecated use {@link PeepholePermalink#LAST_UNSTABLE_BUILD} */
+        @Deprecated
+        public static Permalink LAST_UNSTABLE_BUILD;
 
-            @Override
-            public String getId() {
-                return "lastFailedBuild";
-            }
+        /** @deprecated use {@link PeepholePermalink#LAST_UNSUCCESSFUL_BUILD} */
+        @Deprecated
+        public static Permalink LAST_UNSUCCESSFUL_BUILD;
 
-            @Override
-            public boolean apply(Run<?, ?> run) {
-                return !run.isBuilding() && run.getResult() == Result.FAILURE;
-            }
-        };
-
-        public static final Permalink LAST_UNSTABLE_BUILD = new PeepholePermalink() {
-            @Override
-            public String getDisplayName() {
-                return Messages.Permalink_LastUnstableBuild();
-            }
-
-            @Override
-            public String getId() {
-                return "lastUnstableBuild";
-            }
-
-            @Override
-            public boolean apply(Run<?, ?> run) {
-                return !run.isBuilding() && run.getResult() == Result.UNSTABLE;
-            }
-        };
-
-        public static final Permalink LAST_UNSUCCESSFUL_BUILD = new PeepholePermalink() {
-            @Override
-            public String getDisplayName() {
-                return Messages.Permalink_LastUnsuccessfulBuild();
-            }
-
-            @Override
-            public String getId() {
-                return "lastUnsuccessfulBuild";
-            }
-
-            @Override
-            public boolean apply(Run<?, ?> run) {
-                return !run.isBuilding() && run.getResult() != Result.SUCCESS;
-            }
-        };
-        public static final Permalink LAST_COMPLETED_BUILD = new PeepholePermalink() {
-            @Override
-            public String getDisplayName() {
-                return Messages.Permalink_LastCompletedBuild();
-            }
-
-            @Override
-            public String getId() {
-                return "lastCompletedBuild";
-            }
-
-            @Override
-            public boolean apply(Run<?, ?> run) {
-                return !run.isBuilding();
-            }
-        };
+        /** @deprecated use {@link PeepholePermalink#LAST_COMPLETED_BUILD} */
+        @Deprecated
+        public static Permalink LAST_COMPLETED_BUILD;
 
         static {
             BUILTIN.add(LAST_BUILD);
-            BUILTIN.add(LAST_STABLE_BUILD);
-            BUILTIN.add(LAST_SUCCESSFUL_BUILD);
-            BUILTIN.add(LAST_FAILED_BUILD);
-            BUILTIN.add(LAST_UNSTABLE_BUILD);
-            BUILTIN.add(LAST_UNSUCCESSFUL_BUILD);
-            BUILTIN.add(LAST_COMPLETED_BUILD);
         }
     }
 }
